Added utils class for helpers functions
