+++ backend/app/api/routes/stats.py
@@ -2,10 +2,10 @@ from fastapi import APIRouter, HTTPException, Depends, Query
 from sqlalchemy import text
 from sqlmodel import Session
 from datetime import datetime, timedelta
-from typing import Optional
 
 from app.core.db import get_db
 from app.api.schemas import IssuesOpenClosedResponse, ErrorResponse, DataQualityResponse, IssueFirstResponseTimeResponse, PrSuccessRateResponse, PrAvgClosingTimeResponse, BugResolutionTimeResponse, PrReviewTimeResponse
+from app.core.utils import format_time_delta, format_time_difference
 
 router = APIRouter(prefix="/stats", tags=["stats"])
 
@@ -184,30 +184,6 @@ def get_open_closed_issues(
         )
     
 
-def format_time_difference(seconds):
-    """Format a time difference in seconds to a human-readable string."""
-    if not seconds:
-        return "Unknown"
-        
-    seconds = int(seconds)
-    days, remainder = divmod(seconds, 86400)
-    hours, remainder = divmod(remainder, 3600)
-    minutes, seconds = divmod(remainder, 60)
-    
-    parts = []
-    if days > 0:
-        parts.append(f"{days} day{'s' if days != 1 else ''}")
-    if hours > 0:
-        parts.append(f"{hours} hour{'s' if hours != 1 else ''}")
-    if minutes > 0:
-        parts.append(f"{minutes} minute{'s' if minutes != 1 else ''}")
-    if seconds > 0 and not parts:  # Only show seconds if less than a minute
-        parts.append(f"{seconds} second{'s' if seconds != 1 else ''}")
-        
-    if not parts:
-        return "Just now"
-        
-    return ", ".join(parts)
 
 @router.get(
     "/issues/first-response-time",
@@ -300,24 +276,6 @@ def get_first_response_time(
             detail=f"Error calculating first response time: {str(e)}"
         )
 
-
-def format_time_delta(delta: timedelta) -> str:
-    """Convert timedelta to human-readable string (e.g., '2 days 3 hours')"""
-    seconds = delta.total_seconds()
-    periods = [
-        ('day', 86400),
-        ('hour', 3600),
-        ('minute', 60),
-        ('second', 1)
-    ]
-    
-    parts = []
-    for period_name, period_seconds in periods:
-        if seconds >= period_seconds:
-            period_value, seconds = divmod(seconds, period_seconds)
-            parts.append(f"{int(period_value)} {period_name}{'s' if period_value != 1 else ''}")
-    
-    return " ".join(parts[:2]) if parts else "0 seconds"
 
 
 @router.get(
